Avoid hardcoded type for multinomial

+++ sklearn/linear_model/logistic.py
@@ -1287,7 +1287,7 @@ class LogisticRegression(BaseEstimator, LinearClassifierMixin,
         self.n_iter_ = np.asarray(n_iter_, dtype=np.int32)[:, 0]
 
         if self.multi_class == 'multinomial':
-            self.coef_ = fold_coefs_[0][0].astype(np.float32)
+            self.coef_ = fold_coefs_[0][0].astype(_dtype)
         else:
             self.coef_ = np.asarray(fold_coefs_, dtype=_dtype)
             self.coef_ = self.coef_.reshape(n_classes, n_features +
